spotify/requirements.txt
Flask
requests
spotipy
python-dotenv
supabase
flask-cors
cryptography
click
